(registration-react) collect dataset api and update references
