--- applications/registration-react/src/api/get-dataset-by-uri.js
@@ -1,11 +1,0 @@
-import axios from 'axios';
-
-export const getDatasetByURI = async uri => {
-  const url = `/search-api/datasets/byuri?uri=${uri}`;
-
-  const response = await axios
-    .get(url)
-    .catch(e => console.log(JSON.stringify(e))); // eslint-disable-line no-console
-
-  return response && response.data;
-};
